[contracts/src.ts] rename a few things
