+++ modules/contracts/src.ts/actions/index.ts
@@ -1,5 +1,4 @@
-export { deployContracts } from "./contracts";
+export { deployContracts } from "./deployContracts";
 export { fund, fundCommand } from "./fund";
-export { migrate, migrateCommand } from "./migrate";
 export { newToken, newTokenCommand } from "./newToken";
 export { registerTransfer, registerTransferCommand } from "./registerTransfer";
