+++ modules/contracts/src.ts/tests/migrate.spec.ts
@@ -1,8 +1,7 @@
 import { expect } from "@connext/vector-utils";
 
+import { migrate } from "../migrate";
 import { addressBookPath, alice } from "../tests";
-
-import { migrate } from "./migrate";
 
 describe("migrate()", () => {
   it("should run without error", async () => {
